chore(lint): forge fmt

+++ contracts/src/misc/IWETH.sol
@@ -3,8 +3,8 @@
 pragma solidity >=0.8.28;
 
 interface IWETH {
-    function approve(address, uint) external returns (bool);
+    function approve(address, uint256) external returns (bool);
     function deposit() external payable;
-    function transfer(address to, uint value) external returns (bool);
-    function withdraw(uint) external;
+    function transfer(address to, uint256 value) external returns (bool);
+    function withdraw(uint256) external;
 }
